+++ ella-pos/api/inventory/shopee_push_stock.php
@@ -92,7 +92,7 @@ try {
         $stockItem = [
             'seller_stock' => [
                 [
-                    'stock' => (int) $job['new_quantity']
+                    'stock' => (float) $job['new_quantity']
                 ]
             ]
         ];
